stream.campaign required
