+++ src/click/schema/direct-url/redirect-type-factory.ts
@@ -12,10 +12,7 @@ import {
   WithoutRefererRedirectType,
 } from './type'
 import { RedirectType, StreamResponse } from '../../types'
-import {
-  StreamDirectUrl,
-  StreamRedirectType,
-} from '@/campaign/entity/stream.entity'
+import { StreamDirectUrl, StreamRedirectType } from '@/campaign/types'
 
 @Injectable()
 export class RedirectTypeFactory {
