+++ test/click/filter/filter-all.e2e-spec.ts
@@ -2,7 +2,7 @@ import { INestApplication } from '@nestjs/common'
 import { DataSource } from 'typeorm'
 import { createAuthUser } from '../../utils/helpers'
 import { CampaignBuilder } from '../../utils/entity-builder/campaign-builder'
-import { StreamActionType } from '@/campaign/entity/stream.entity'
+import { StreamActionType } from '@/campaign/types'
 import { FilterLogic, FilterObject } from '@/stream-filter/types'
 import { ClickActionBuilder } from '../../utils/click-action-builder'
 import { truncateTables } from '../../utils/truncate-tables'
