+++ test/click/filter/filter-click-limit.e2e-spec.ts
@@ -2,7 +2,7 @@ import { INestApplication } from '@nestjs/common'
 import { DataSource } from 'typeorm'
 import { createAuthUser } from '../../utils/helpers'
 import { CampaignBuilder } from '../../utils/entity-builder/campaign-builder'
-import { StreamActionType } from '@/campaign/entity/stream.entity'
+import { StreamActionType } from '@/campaign/types'
 import { FilterLogic } from '@/stream-filter/types'
 import { ClickActionBuilder } from '../../utils/click-action-builder'
 import { DateTime } from 'luxon'
